FaceLandmarkImg with the new interface.

+++ lib/local/GazeAnalyser/src/GazeEstimation.cpp
@@ -135,7 +135,7 @@ void GazeAnalysis::EstimateGaze(const LandmarkDetector::CLNF& clnf_model, cv::Po
 	gaze_absolute = gazeVecAxis / norm(gazeVecAxis);
 }
 
-cv::Vec2d GazeAnalysis::GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& gaze_vector_2)
+cv::Vec2f GazeAnalysis::GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& gaze_vector_2)
 {
 
 	cv::Point3f gaze_vector = (gaze_vector_1 + gaze_vector_2) / 2;
@@ -143,6 +143,6 @@ cv::Vec2d GazeAnalysis::GetGazeAngle(cv::Point3f& gaze_vector_1, cv::Point3f& ga
 	double x_angle = atan2(gaze_vector.x, -gaze_vector.z);
 	double y_angle = atan2(gaze_vector.y, -gaze_vector.z);
 
-	return cv::Vec2d(x_angle, y_angle);
+	return cv::Vec2f(x_angle, y_angle);
 
 }
